Move the PSP system language logic into sceUtility.
